tests: Split buffer property tests by in/out stream
